CB-3865 Workspace based authorization not working for periscope
